requirements.txt
django
djangorestframework
psycopg2-binary
gunicorn
django-filter
django-cors-headers
python-dotenv
djangorestframework-simplejwt
django-jazzmin
drf-spectacular
stripe
dj-database-url
boto3
django-storages
